Prototype work on background web requests and monitoring,  primarily for better exception handling,  using the "futures" package.   Still a black hole...
