+++ crds/web.py
@@ -8,7 +8,7 @@ import threading
 import requests
 from lxml import html
 
-from crds import config, log, utils, exceptions
+from crds import config, log, utils, exceptions, background
 from crds.python23 import *
 
 # ==================================================================================================
@@ -18,32 +18,6 @@ def log_section(section_name, section_value, verbosity=50, log_function=log.verb
     """Issue log divider bar followed by a corresponding log message."""
     log.divider(name=divider_name, verbosity=verbosity)
     log_function(section_name, section_value, verbosity=verbosity+5)
-
-def background(f):
-    """a threading decorator use @background above the function you want to run in the background.
-    The decorated function returns (thread, queue) where queue will contain the function result
-    and thead is already started but not joined.
-    """
-
-    def run_thread(*args, **keys):
-        q = queue.Queue()
-        def queue_put_f():
-            q.put(f(*args, **keys))
-        t = threading.Thread(target=queue_put_f)
-        t.start()
-        return t, q
-
-    run_thread.__name__ = f.__name__ + "[background]"
-
-    return run_thread
-
-def background_complete(args):
-    if isinstance(args, tuple) and len(args) == 2:
-        args[0].join()
-        return args[1].get()
-    else:
-        response = args
-    return response
 
 # ==================================================================================================
 
@@ -79,7 +53,7 @@ class CrdsDjangoConnection(object):
         log.divider()
 
     def response_complete(self, args):
-        response = background_complete(args)
+        response = background.background_complete(args)
         self.dump_response("Response: ", response)
         self.check_error(response)
         return response
@@ -91,7 +65,7 @@ class CrdsDjangoConnection(object):
         args = self.get_start(relative_url)
         return self.get_complete(args)
     
-    @background
+    @background.background
     def get_start(self, relative_url):
         url = self.abs_url(relative_url)
         log_section("GET:", url, divider_name="GET: " + url.split("&")[0])
@@ -102,7 +76,7 @@ class CrdsDjangoConnection(object):
         args = self.post_start(relative_url, *post_dicts, **post_vars)
         return self.post_complete(args)
 
-    @background
+    @background.background
     def post_start(self, relative_url, *post_dicts, **post_vars):
         url = self.abs_url(relative_url)
         vars = utils.combine_dicts(*post_dicts, **post_vars)
